+++ src/main/scala/millfork/compiler/ExpressionCompiler.scala
@@ -386,7 +386,7 @@ object ExpressionCompiler {
 
   def callingContext(ctx: CompilationContext, v: MemoryVariable): CompilationContext = {
     val result = new Environment(Some(ctx.env), "")
-    result.registerVariable(VariableDeclarationStatement(v.name, v.typ.name, stack = false, global = false, constant = false, volatile = false, register = false, initialValue = None, address = None), ctx.options)
+    result.registerVariable(VariableDeclarationStatement(v.name, v.typ.name, stack = false, global = false, constant = false, volatile = false, register = false, initialValue = None, address = None, bank = v.declaredBank), ctx.options)
     ctx.copy(env = result)
   }