+++ src/test/scala/millfork/test/emu/EmuPlatform.scala
@@ -1,7 +1,7 @@
 package millfork.test.emu
 
 import millfork.output.{AfterCodeByteAllocator, CurrentBankFragmentOutput, UpwardByteAllocator, VariableAllocator}
-import millfork.{Cpu, Platform}
+import millfork.{Cpu, OutputStyle, Platform}
 
 /**
   * @author Karol Stasiak
@@ -12,8 +12,11 @@ object EmuPlatform {
     Map(),
     Nil,
     CurrentBankFragmentOutput(0, 0xffff),
-    new UpwardByteAllocator(0x200, 0xb000),
-    new VariableAllocator((0 until 256 by 2).toList, new AfterCodeByteAllocator(0xff00)),
-    ".bin"
+    Map("default" -> new UpwardByteAllocator(0x200, 0xb000)),
+    Map("default" -> new VariableAllocator((0 until 256 by 2).toList, new AfterCodeByteAllocator(0xff00))),
+    ".bin",
+    Map("default" -> 0),
+    "default",
+    OutputStyle.Single
   )
 }